major reorganization of geometries header files
