+++ inc/geometries/README
@@ -1,0 +1,43 @@
+
+#grids
+    conformal.h
+    orthogonal.h
+    curvilinear.h
+    mpi_conformal.h
+    mpi_orthogonal.h
+    mpi_curvilinear.h
+    refined_conformal.h
+    refined_orthogonal.h
+    refined_curvilinear.h
+
+    orthogonalX.h
+    curvilinearX.h
+    refined_orthogonalX.h
+
+#generators
+    orthogonal_generators.h
+    orthogonalX_generators.h
+    ribeiro.h
+    ribeiroX.h
+    flux.h
+    hector.h
+
+    # utilities
+        utilities.h
+        utilitiesX.h
+
+#magnetic fields
+    solovev.h
+    solovev_parameters.h
+    taylor.h
+    guenther.h
+    guenther_parameters.h
+
+    testfunctors.h 
+    init.h
+    magnetic_field.h
+    adaption.h
+
+#miscellaneous
+    average.h
+
